requirements.txt
google-generativeai
python-dotenv
pandas
scikit-learn
chromadb
pypdf
matplotlib
statsmodels
reportlab
